+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v3.9.0 [#](https://github.com/idleberg/vscode-nsis/releases/tag/3.9.0)
+
+- add `showVersionAsInfoMessage` option
+- add `showFlagsAsObject` option
+
 # v3.8.0 [#](https://github.com/idleberg/vscode-nsis/releases/tag/3.8.0)
 
 - make use of `makensis` module